pyproject.toml
[project]
name = "finsight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.28.8",
    "openai>=1.12.0",
    "pandas>=2.0.0",
    "pathlib2>=2.3.0",
    "pillow>=11.3.0",
    "pymupdf>=1.23.8",
    "python-dotenv>=1.0.0",
]
